Solo me falta un fallo en la programación dinámica y la documentación

+++ p06/src/TSP/programacion_dinamica_tsp.h
@@ -12,7 +12,7 @@ class ProgramacionDinamicaTSP: public TSP {
   virtual void mostrarTiempoEjecución() const override;
  protected:
   void reconstruirCaminoMinimo(
-    const std::vector<std::vector<long double>>&,
+    const std::vector<std::vector<int>>&,
     const GrafoNoDirigidoCompleto&
   );
 };
